Factor the remove-instanceof path out of remove-virtual-functions.
