+++ src/goto-programs/remove_instanceof.h
@@ -1,0 +1,13 @@
+
+#ifndef REMOVE_INSTANCEOF_H
+#define REMOVE_INSTANCEOF_H
+
+#include <util/symbol_table.h>
+#include "goto_functions.h"
+
+void remove_instanceof(
+  symbol_tablet &symbol_table,
+  goto_functionst &goto_functions);
+
+
+#endif
